refactor: split 11 oversized components under 250-line rule
Extract pure logic, helpers, types, and presentational sub-components
into co-located sibling files, keeping each component as a thin
orchestrator. No behavior or rendered output changed.

- NotificationCenter 497->292 (+notificationData.ts)
- QuickInsights 421->226 (+quickInsightsData.ts)
- RecurringTransactions 361->246 (+recurringUtils.ts)
- CategoryBreakdown 352->270 (+categoryBreakdownUtils.ts)
- TransactionTable 339->241 (+transactionColumns.tsx)
- BudgetTracker 333->263 (+budgetUtils.ts, +components/AddBudgetForm.tsx)
- Sparkline 324->230 (+sparklineUtils.ts, +SparklineCompact.tsx)
- CashFlowForecast 305->160 (+cashFlowUtils.ts)
- PeriodComparison 304->248 (+periodComparisonUtils.ts)
- ProfileModal 274->249 (+profileModalUtils.ts, +profile-modal/LogoutButton.tsx)
- InstrumentProjections 274->186 (+instrumentProjectionUtils.ts, +instrumentProjectionComponents.tsx)

frontend 190 tests + tsc + eslint green.
