+++ frontend/src/components/analytics/InstrumentProjections.tsx
@@ -1,96 +1,23 @@
 import { useState, useMemo } from 'react'
 import { motion } from 'framer-motion'
 import { Landmark, IndianRupee, PiggyBank, TrendingUp, Percent } from 'lucide-react'
-import { rawColors } from '@/constants/colors'
 import { formatCurrency } from '@/lib/formatters'
 import { StaleDataBadge } from '@/components/shared/StaleDataBadge'
-import StandardAreaChart from '@/components/analytics/StandardAreaChart'
 import MetricCard from '@/components/shared/MetricCard'
-import {
-  projectPPF,
-  projectEPF,
-  projectNPS,
-  EPF_MIN_MONTHLY_CONTRIBUTION,
-  EPF_STATUTORY_RATE_PCT,
-} from '@/lib/instrumentCalculators'
-import type { ProjectionResult } from '@/lib/instrumentCalculators'
+import { projectPPF, projectEPF, projectNPS } from '@/lib/instrumentCalculators'
 import { useAccountBalances } from '@/hooks/api/useAnalytics'
 import { useInstrumentRates } from '@/hooks/api/useInstrumentRates'
 import type { InstrumentRates } from '@/services/api/rates'
-import type { AccountBalances } from '@/services/api/calculations'
 
-type Instrument = 'ppf' | 'epf' | 'nps'
-
-const TABS: { key: Instrument; label: string }[] = [
-  { key: 'ppf', label: 'PPF' },
-  { key: 'epf', label: 'EPF' },
-  { key: 'nps', label: 'NPS' },
-]
-
-function SliderInput({
-  id,
-  label,
-  value,
-  onChange,
-  min,
-  max,
-  step,
-  suffix,
-  prefix,
-}: Readonly<{
-  id: string
-  label: string
-  value: number
-  onChange: (v: number) => void
-  min: number
-  max: number
-  step: number
-  suffix?: string
-  prefix?: string
-}>) {
-  return (
-    <div>
-      <div className="flex items-center justify-between mb-1.5">
-        <label htmlFor={id} className="text-xs text-text-secondary">{label}</label>
-        <span className="text-sm font-medium text-white">
-          {prefix}{value.toLocaleString('en-IN')}{suffix}
-        </span>
-      </div>
-      <input
-        id={id}
-        type="range"
-        min={min}
-        max={max}
-        step={step}
-        value={value}
-        onChange={(e) => onChange(Number(e.target.value))}
-        className="w-full h-1.5 rounded-full appearance-none bg-white/10 accent-app-blue cursor-pointer"
-      />
-    </div>
-  )
-}
-
-function ProjectionChart({ data }: Readonly<{ data: ProjectionResult }>) {
-  const chartData = data.yearByYear.map((y) => ({
-    year: `Y${y.year}`,
-    Contributed: y.contributed,
-    Returns: y.returns,
-  }))
-
-  return (
-    <StandardAreaChart
-      data={chartData}
-      dataKey="year"
-      height={280}
-      stacked
-      tooltipFormatter={formatCurrency}
-      areas={[
-        { key: 'Contributed', color: rawColors.app.blue, fillOpacity: 0.7 },
-        { key: 'Returns', color: rawColors.app.green, fillOpacity: 0.7 },
-      ]}
-    />
-  )
-}
+import { ProjectionChart, SliderInput } from './instrumentProjectionComponents'
+import {
+  TABS,
+  computeEpfContribution,
+  computeWeightedReturn,
+  findAccountBalance,
+  rebalanceNpsAllocation,
+  type Instrument,
+} from './instrumentProjectionUtils'
 
 function PPFTab({
   initialBalance,
@@ -133,16 +60,12 @@ function EPFTab({
   const [years, setYears] = useState(25)
   const [balance, setBalance] = useState(initialBalance)
 
-  // Employee + employer both contribute the same %, total = 2x
-  const yourShare = salary * contribPct / 100
-  const totalMonthly = yourShare * 2
-  // Floor is the statutory minimum; above the wage ceiling you contribute on full salary.
-  const minContrib = Math.max(EPF_MIN_MONTHLY_CONTRIBUTION, (salary * EPF_STATUTORY_RATE_PCT) / 100)
-
   const result = useMemo(
     () => projectEPF(salary, contribPct, contribPct, rate, years, balance),
     [salary, contribPct, rate, years, balance],
   )
+
+  const { yourShare, totalMonthly, minContrib } = computeEpfContribution(salary, contribPct)
 
   return (
     <div className="space-y-5">
@@ -176,13 +99,11 @@ function NPSTab({ rates }: Readonly<{ rates: InstrumentRates }>) {
   const [years, setYears] = useState(25)
   const [balance, setBalance] = useState(0)
 
-  // Keep allocation summing to 100
   const handleEquity = (v: number) => {
-    const remaining = 100 - v
-    const ratio = corp + govt > 0 ? corp / (corp + govt) : 0.5
-    setEquity(v)
-    setCorp(Math.round(remaining * ratio))
-    setGovt(remaining - Math.round(remaining * ratio))
+    const next = rebalanceNpsAllocation(v, corp, govt)
+    setEquity(next.equity)
+    setCorp(next.corp)
+    setGovt(next.govt)
   }
 
   const result = useMemo(() => projectNPS({
@@ -196,10 +117,7 @@ function NPSTab({ rates }: Readonly<{ rates: InstrumentRates }>) {
     years,
     currentBalance: balance,
   }), [monthly, equity, corp, govt, returns, years, balance])
-  const weightedReturn =
-    (equity / 100) * returns.equity +
-    (corp / 100) * returns.corp_bond +
-    (govt / 100) * returns.govt_bond
+  const weightedReturn = computeWeightedReturn(equity, corp, govt, returns)
 
   return (
     <div className="space-y-5">
@@ -222,12 +140,6 @@ function NPSTab({ rates }: Readonly<{ rates: InstrumentRates }>) {
       <p className="text-xs text-text-tertiary">NPS Tier-I: Max 75% equity (auto-choice). At maturity, 60% is tax-free lump sum, 40% must buy annuity. Extra Rs 50K deduction under 80CCD(1B).</p>
     </div>
   )
-}
-
-function findAccountBalance(data: AccountBalances | undefined, pattern: string): number {
-  if (!data?.accounts) return 0
-  const key = Object.keys(data.accounts).find((k) => k.toLowerCase().includes(pattern))
-  return key ? Math.max(0, data.accounts[key].balance) : 0
 }
 
 export default function InstrumentProjections() {
