+++ frontend/src/components/shared/Sparkline.tsx
@@ -5,6 +5,9 @@ import { motion, useMotionValue, useTransform, animate } from 'framer-motion'
 import { rawColors } from '@/constants/colors'
 import { formatCurrencyShort } from '@/lib/formatters'
 
+import CompactSparkline from './SparklineCompact'
+import { buildDefaultGeometry } from './sparklineUtils'
+
 interface SparklineProps {
   data: number[]
   color?: string
@@ -21,75 +24,6 @@ interface SparklineProps {
   width?: number
   /** Optional aria-label. Compact callsites should always provide one. */
   ariaLabel?: string
-}
-
-const COMPACT_VIEWBOX_W = 100
-const COMPACT_VIEWBOX_H = 30
-
-/**
- * Compact-variant sparkline: pure SVG, no chart-lib overhead, no
- * animation/hover. Mirrors the inline-list use case (category breakdown
- * rows, table cells). Falls back to the rich animated form when
- * variant is omitted or 'default'.
- */
-function CompactSparkline({
-  data,
-  color,
-  width,
-  height,
-  ariaLabel,
-}: Readonly<{
-  data: number[]
-  color: string
-  width: number
-  height: number
-  ariaLabel?: string
-}>) {
-  if (data.length < 2) return null
-
-  const max = Math.max(...data)
-  const min = Math.min(...data)
-  // Flat series: avoid divide-by-zero, render a horizontal mid-line.
-  const span = max - min === 0 ? 1 : max - min
-  const stepX = COMPACT_VIEWBOX_W / (data.length - 1)
-
-  const points = data.map((v, i) => {
-    const x = i * stepX
-    // Y inverted (SVG top-left origin); 2 px top + bottom padding.
-    const y =
-      COMPACT_VIEWBOX_H - 2 - ((v - min) / span) * (COMPACT_VIEWBOX_H - 4)
-    return { x, y }
-  })
-
-  const linePath = points
-    .map((p, i) => `${i === 0 ? 'M' : 'L'} ${p.x.toFixed(2)} ${p.y.toFixed(2)}`)
-    .join(' ')
-  const areaPath = `${linePath} L ${COMPACT_VIEWBOX_W} ${COMPACT_VIEWBOX_H} L 0 ${COMPACT_VIEWBOX_H} Z`
-
-  const last = points[points.length - 1]
-
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox={`0 0 ${COMPACT_VIEWBOX_W} ${COMPACT_VIEWBOX_H}`}
-      preserveAspectRatio="none"
-      role="img"
-      aria-label={ariaLabel ?? 'Trend'}
-      className="shrink-0 overflow-visible"
-    >
-      <path d={areaPath} fill={color} fillOpacity={0.15} />
-      <path
-        d={linePath}
-        fill="none"
-        stroke={color}
-        strokeWidth={1.5}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <circle cx={last.x} cy={last.y} r={1.6} fill={color} />
-    </svg>
-  )
 }
 
 /**
@@ -148,38 +82,10 @@ function DefaultSparkline({
   const width = 200
   const [hoverIndex, setHoverIndex] = useState<number | null>(null)
 
-  const { linePath, areaPath, points, avgY } = useMemo(() => {
-    if (data.length < 2) return { linePath: '', areaPath: '', points: [], avgY: 0 }
-
-    const min = Math.min(...data)
-    const max = Math.max(...data)
-    const range = max - min || 1
-    const padding = 4
-
-    const pts = data.map((v, i) => {
-      const x = (i / (data.length - 1)) * width
-      const y = padding + ((max - v) / range) * (height - padding * 2)
-      return { x, y, value: v }
-    })
-
-    let line = `M ${pts[0].x},${pts[0].y}`
-    for (let i = 1; i < pts.length; i++) {
-      const prev = pts[i - 1]
-      const curr = pts[i]
-      const cpx = (prev.x + curr.x) / 2
-      line += ` C ${cpx},${prev.y} ${cpx},${curr.y} ${curr.x},${curr.y}`
-    }
-
-    const last = pts.at(-1)
-    if (!last) return { linePath: '', areaPath: '', points: [], avgY: 0 }
-    const area = `${line} L ${last.x},${height} L ${pts[0].x},${height} Z`
-
-    // Average value and its Y coordinate for the reference line
-    const avg = data.reduce((sum, v) => sum + v, 0) / data.length
-    const averageY = padding + ((max - avg) / range) * (height - padding * 2)
-
-    return { linePath: line, areaPath: area, points: pts, avgY: averageY }
-  }, [data, height])
+  const { linePath, areaPath, points, avgY } = useMemo(
+    () => buildDefaultGeometry(data, width, height),
+    [data, height],
+  )
 
   const progress = useMotionValue(0)
